separate wallet module
